requirements.txt
asgiref==3.2.10
boto3==1.17.24
botocore==1.20.24
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.43.0
django-appconf==1.0.4
django-imagekit==4.0.2
django-import-export==2.5.0
django-qr-code==1.2.0
django-storages==1.11.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
jmespath==0.10.0
MarkupPy==1.14
numpy==1.19.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
pilkit==2.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
twilio==6.45.1
urllib3==1.25.10
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
 